refactor(chat): update Chat and ChatMessage interfaces and remove mock conversations
