+++ src/@types/conversation.entity.ts
@@ -27,3 +27,22 @@ export interface ChatSession {
   conversation: Conversation;
   messages: ChatMessage[];
 }
+
+export interface Chat {
+  id: string;
+  participants: string[];
+  lastMessage: string;
+  updatedAt: any;
+}
+
+export interface ChatMessage {
+  id: string;
+  senderId: string;
+  text: string;
+  createdAt: any;
+  attachment?: {
+    name: string;
+    type: string;
+    data: string;
+  };
+}
